pyproject.toml
[project]
name = "dj"
version = "0.1.0"
description = ""
authors = [
    {name = "ArBes"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
